Move image crop logic to util
