+++ origin-dapp/src/components/modals/image-cropper.js
@@ -1,11 +1,8 @@
 import React, { Component, Fragment } from 'react'
 import { FormattedMessage } from 'react-intl'
 import ReactCrop from 'react-image-crop'
-import { getDataUri } from 'utils/fileUtils'
+import { getDataUri, generateCroppedImage } from 'utils/fileUtils'
 import Modal from 'components/modal'
-
-const MAX_IMAGE_WIDTH = 1000
-const MAX_IMAGE_HEIGHT = 1000
 
 class ImageCropper extends Component {
   constructor(props) {
@@ -28,7 +25,6 @@ class ImageCropper extends Component {
 
     this.onCropChange = this.onCropChange.bind(this)
     this.onCropComplete = this.onCropComplete.bind(this)
-    this.generateCroppedImage = this.generateCroppedImage.bind(this)
   }
 
   async componentDidMount() {
@@ -47,70 +43,10 @@ class ImageCropper extends Component {
   }
 
   async onCropComplete() {
-    const croppedImageFile = await this.generateCroppedImage()
+    const { imageSrc, imageFileObj, pixelCrop } = this.state
+    const croppedImageFile = await generateCroppedImage(imageSrc, imageFileObj, pixelCrop)
     const croppedImageUri = await getDataUri(croppedImageFile)
     this.props.onCropComplete(croppedImageUri)
-  }
-
-  generateCroppedImage() { 
-    const { pixelCrop, imageFileObj, imageSrc } = this.state
-    let image
-    let canvas
-    
-    function drawImageOnCanvas(imgEl) {
-      const defaultConfig = {
-        x: 0,
-        y: 0,
-        width: imgEl.width,
-        height: imgEl.height
-      }
-
-      const { x, y, width, height } = pixelCrop || defaultConfig
-
-      let resizedWidth = width
-      let resizedHeight = height
-
-      if (width > MAX_IMAGE_WIDTH) {
-        resizedWidth = MAX_IMAGE_WIDTH
-        const widthDiffRatio = resizedWidth / width
-        resizedHeight = height * widthDiffRatio
-      }
-
-      if (resizedHeight > MAX_IMAGE_HEIGHT) {
-        const heightDiffRatio = MAX_IMAGE_HEIGHT / resizedHeight
-        resizedHeight = MAX_IMAGE_HEIGHT
-        resizedWidth = resizedWidth * heightDiffRatio
-      }
-
-      canvas = document.createElement('canvas')
-      canvas.width = resizedWidth
-      canvas.height = resizedHeight
-      const ctx = canvas.getContext('2d')
-
-      ctx.drawImage(
-        image,
-        x,
-        y,
-        width,
-        height,
-        0,
-        0,
-        resizedWidth,
-        resizedHeight
-      )
-    }
-   
-    return new Promise((resolve) => {
-      image = new Image()
-      image.onload = () => {
-        drawImageOnCanvas(image)
-        canvas.toBlob(file => {
-          file.name = imageFileObj.name
-          resolve(file)
-        }, 'image/jpeg')
-      }
-      image.src = imageSrc
-    })
   }
 
   render() {
